separation des classes en fichiers distincts
