+++ assets/js/theme/components/carousel/slider.js
@@ -120,76 +120,3 @@ window.osuny.carousel.Slider.prototype = {
         return slideIndex;
     }
 }
-
-window.osuny.carousel.Slide = function (slider, container, i) {
-    this.size;
-    this.container = container;
-    this.slider = slider;
-    this.initialize();
-    this.index = i;
-}
-
-window.osuny.carousel.Slide.prototype = {
-    initialize: function () {
-        this.computeSize();
-    },
-    computeSize: function () {
-        var style = getComputedStyle(this.container);
-        this.size = this.container.offsetWidth + parseFloat(style.marginLeft) + parseFloat(style.marginRight);
-    },
-    setClasses() {
-        this.cleanStateClasses();
-        if (this.slider.index == this.index) {
-            this.setCurrent();
-        }
-        if (this.slider.index - 1 == this.index) {
-            this.setPrevious();
-        }
-        if (this.slider.index > this.index) {
-            this.setBefore();
-        }
-        if (this.slider.index + 1 == this.index) {
-            this.setNext();
-        }
-    },
-    setCurrent: function () {
-        this.container.classList.add('is-current');
-    },
-    setPrevious: function () {
-        this.container.classList.add('is-previous');
-    },
-    setBefore: function () {
-        this.container.classList.add('is-before');
-    },
-    setNext: function () {
-        this.container.classList.add('is-next');
-    },
-    cleanStateClasses: function () {
-        this.container.classList.remove('is-current');
-        this.container.classList.remove('is-next');
-        this.container.classList.remove('is-previous');
-        this.container.classList.remove('is-before');
-    }
-}
-
-window.osuny.carousel.Drag = function (slider) {
-    this.active = false;
-    this.initialPosition = 0;
-    this.slider = slider;
-}
-window.osuny.carousel.Drag.prototype = {
-    start: function (position) {
-        this.active = true;
-        this.initialPosition = position;
-    },
-    drag: function (position) {
-        var distance = this.initialPosition - position;
-        this.slider.deltaPosition -= distance;
-        this.slider.translate();
-    },
-    end: function () {
-        this.active = false;
-        this.initialPosition = 0;
-        this.slider.recomputePosition();
-    }
-}
